Greedy now keeps trying different ics
Instead of using only the cheapest one in terms of price per requests,
the greedy allocator now tries all the instance classes in order. The
problem is considered only infeasible after all ics have been tried.

One test has to be rewritten, since it was assuming a non feasible
problem, which is now feasible. And a new test was added with
an infeasible problem for the current algorithm.