Make Greedy use different ICs, and add poetry configuration
Instead of using only the cheapest one in terms of price per requests,
the greedy allocator now tries all the instance classes in order. The
problem is considered only infeasible after all ics have been tried.

One test has to be rewritten, since it was assuming a non-feasible
problem, which is now feasible. A new test was added with
an infeasible problem for the current algorithm.

Also added small example for greedy allocator.

In addition, poetry configuration was added and a sum of requests was added to
solution visualization.