Optimization of the hash of ICs and CCs
The creation of the problem spent most of the time in the hash
function of instance classes and container classes, because they
have pint objects that spend a lot of time in the hash function.
Now, only the name is used to hash the ICs and CCs. A check was
added in the system creation to ensure that the names of the ICs
and CCs are unique.